Separate translators for building and stripping HTML.
